+++ example/lib/src/alignment_flex_table.dart
@@ -30,8 +30,8 @@ class _AlignmentFlexTableState extends State<AlignmentFlexTable> {
   late final FlexSortController _sortController =
       FlexSortController(controller: _controller);
 
-  final List<String> _headers = [
-    for (var i = 1; i <= kColumns; i++) 'Random #$i'
+  final List<FlexTableCell> _headers = [
+    for (var i = 1; i <= kColumns; i++) FlexTableCell(value: 'Random #$i')
   ];
 
   late final Timer _timer;
@@ -40,12 +40,12 @@ class _AlignmentFlexTableState extends State<AlignmentFlexTable> {
   void initState() {
     super.initState();
     _timer = Timer.periodic(const Duration(milliseconds: 100), (_) {
-      final columns = <String>[];
+      final columns = <FlexTableCell>[];
       final random = Random();
 
       for (var i = 0; i < _headers.length; i++) {
         final word = nouns[random.nextInt(nouns.length)];
-        columns.add(word);
+        columns.add(FlexTableCell(value: word));
       }
 
       _controller.addRow(FlexTableRow(columns: columns));