+++ example/lib/src/dynamic_flex_table.dart
@@ -29,8 +29,8 @@ class _DynamicFlexTableState extends State<DynamicFlexTable> {
   late final FlexSortController _sortController =
       FlexSortController(controller: _controller);
 
-  final List<String> _headers = [
-    for (var i = 1; i <= kColumns; i++) 'Random #$i'
+  final List<FlexTableCell> _headers = [
+    for (var i = 1; i <= kColumns; i++) FlexTableCell(value: 'Random #$i')
   ];
 
   late final Timer _timer;
@@ -39,12 +39,12 @@ class _DynamicFlexTableState extends State<DynamicFlexTable> {
   void initState() {
     super.initState();
     _timer = Timer.periodic(const Duration(milliseconds: 100), (_) {
-      final columns = <String>[];
+      final columns = <FlexTableCell>[];
       final random = Random();
 
       for (var i = 0; i < kColumns; i++) {
         final word = nouns[random.nextInt(nouns.length)];
-        columns.add(word);
+        columns.add(FlexTableCell(value: word));
       }
 
       _controller.addRow(FlexTableRow(columns: columns));
@@ -89,6 +89,13 @@ class _DynamicFlexTableState extends State<DynamicFlexTable> {
                 controller: _sortController,
               )
             : const FlexHeaderCellBuilder(),
+        onRowSelected: (index) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text('Row Selected: $index'),
+            ),
+          );
+        },
       ),
     );
   }