+++ example/lib/src/not_resizable_flex_table.dart
@@ -23,8 +23,8 @@ class _NotResizableFlexTableState extends State<NotResizableFlexTable> {
 
   final FlexTableController _controller = FlexTableController();
 
-  final List<String> _headers = [
-    for (var i = 1; i <= kColumns; i++) 'Random #$i'
+  final List<FlexTableCell> _headers = [
+    for (var i = 1; i <= kColumns; i++) FlexTableCell(value: 'Random #$i'),
   ];
 
   late final Timer _timer;
@@ -33,12 +33,12 @@ class _NotResizableFlexTableState extends State<NotResizableFlexTable> {
   void initState() {
     super.initState();
     _timer = Timer.periodic(const Duration(milliseconds: 100), (_) {
-      final columns = <String>[];
+      final columns = <FlexTableCell>[];
       final random = Random();
 
       for (var i = 0; i < kColumns; i++) {
         final word = nouns[random.nextInt(nouns.length)];
-        columns.add(word);
+        columns.add(FlexTableCell(value: word));
       }
 
       _controller.addRow(FlexTableRow(columns: columns));