+++ lib/flexible_table.dart
@@ -1,1 +1,3 @@
 export 'src/flex_table.dart';
+export 'src/flex_table_cell.dart';
+export 'src/flex_table_row.dart';