+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.0] - October 6TH, 2024
+
+* Added ability to use widgets other than Text
+* Added ability to inform table of selected rows.
+* Added ability to get informed when a row is selected.
+* Depends on [provider](https://pub.dev/packages/provider) package now.
+
+
 ## [1.0.0+2] - October 1, 2024
 
 * Automated dependency updates